img_similarity/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asian-bart @ file:///C:/Users/SSAFY/Desktop/%EC%9E%90%EC%9C%A8PJT/S11P31A105/img_similarity/asian-bart
braceexpand==0.1.7
cachetools==5.5.0
certifi==2024.8.30
chardet==3.0.4
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
distlib==0.3.9
exceptiongroup==1.2.2
fastapi==0.115.4
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.10.0
ftfy==6.3.1
google-api-core==2.22.0
google-auth==2.35.0
google-cloud-vision==3.8.0
googleapis-common-protos==1.65.0
googletrans==4.0.0rc1
grpcio==1.67.1
grpcio-status==1.67.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.11.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.26.2
hyperframe==5.2.0
idna==2.10
importlib_resources==6.4.5
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.7.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.24.4
-e git+https://github.com/mlfoundations/open_clip.git@49eac2f27a5bb98a7f7ecc1154918880aa55256c#egg=open_clip_torch
opencv-python==4.10.0.84
packaging==24.1
pandas==2.0.3
pillow==10.4.0
platformdirs==4.3.6
proto-plus==1.25.0
protobuf==5.28.3
psutil==6.1.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-multipart==0.0.17
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rfc3986==1.5.0
rsa==4.9
safetensors==0.4.5
scipy==1.13.1
seaborn==0.13.2
sentencepiece==0.2.0
six==1.16.0
sniffio==1.3.1
starlette==0.41.2
submitit==1.5.2
sympy==1.13.1
timm==1.0.11
tokenizers==0.20.1
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.66.5
transformers==4.46.2
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.3.27
ultralytics-thop==2.0.10
urllib3==2.2.2
uvicorn==0.32.0
virtualenv==20.26.3
wcwidth==0.2.13
zipp==3.20.2
